Update script and remove --tw
