+++ src/pagebuilder/sass/StandardCard.module.scss
@@ -10,8 +10,7 @@
     border-radius: 0.5rem;
     margin-bottom: 1rem;
     padding: 1rem;
-    --tw-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
-    box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
+    box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.05);
   }
 
   &__header {
@@ -57,8 +56,7 @@
     font-weight: 500;
     border-radius: var(--yxt-cta-border-radius);
     padding: 0.5rem 1.25rem;
-    --tw-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06);
-    box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
+    box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.06), 0px 1px 3px rgba(0, 0, 0, 0.1);
   }
 
   &__cta1 {
