+++ src/pagebuilder/sass/answers-variables.scss
@@ -1,6 +1,6 @@
 :root {
-	--yxt-color-brand-primary: red;
-	--yxt-base-font-size: 14px;
-	--yxt-cta-border-radius: 16px;
-	--yxt-cta-color: green;
+  --yxt-color-brand-primary: red;
+  --yxt-base-font-size: 14px;
+  --yxt-cta-border-radius: 16px;
+  --yxt-cta-color: green;
 }
