+++ src/pagebuilder/scripts/generateAnswersVariables.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const chokidar = require('chokidar');
-const createAnswersVariables = require('./createAnswersVariables');
+const createAnswersVariablesFile = require('./createAnswersVariablesFile');
 
 const configPath = __dirname + '/../pageBuilderAnswersAppConfig.json';
 generateAnswersVariables();
@@ -9,5 +9,5 @@ chokidar.watch(configPath).on('change', () => generateAnswersVariables());
 
 function generateAnswersVariables() {
   const config = JSON.parse(fs.readFileSync(configPath));
-  createAnswersVariables(config);
+  createAnswersVariablesFile(config);
 }
